fix(worker_threads): root listeners before MessageEvent alloc; split channel pump; regen API docs
Addresses CI + CodeRabbit feedback on #5531:

- GC-safety: root the listener closures in the handle scope BEFORE allocating
  the `MessageEvent` in both dispatch paths (`deliver_parent_port_message`,
  `dispatch_worker_event`). The allocation can trigger a moving GC, which
  rewrites the canonical listener storage via the registered root scanners;
  a plain snapshot of the raw closure bits would otherwise call relocated
  closures.
- lint: extract the MessageChannel/BroadcastChannel event-loop pump
  (`js_worker_threads_channels_process_pending` / `_has_pending`) into a new
  `worker_threads/channel_pump.rs` submodule, re-exported from the parent so
  `crate::worker_threads::*` paths still resolve. `worker_threads.rs` was 2157
  lines (>2000 cap); now 1957.
- api-docs-drift: regenerate docs/src/api/reference.md for the new
  addEventListener/removeEventListener manifest entries.

Verified: perry-stdlib builds clean; both cross-thread e2e regression tests pass
(worker_pool_all_workers_execute, add_event_listener_round_trip).
