+++ docs/src/api/reference.md
@@ -2,7 +2,7 @@
 
 This page is auto-generated from Perry's compile-time API manifest (`perry-api-manifest::API_MANIFEST`). It is the source of truth for what `perry compile` accepts; references to symbols not listed here produce `R005 UnimplementedApi` (issue #463). Stubs (#464) are flagged ⚠ — they link cleanly but no-op at runtime on the chosen target.
 
-Total: 2817 entries across 115 modules.
+Total: 2819 entries across 115 modules.
 
 ## Modules
 
@@ -3676,6 +3676,7 @@ Total: 2817 entries across 115 modules.
 
 - `BroadcastChannel` — module
 - `MessageChannel` — module
+- `addEventListener` — instance
 - `cpuUsage` — instance *(class: `Worker`)*
 - `getEnvironmentData` — module
 - `getHeapSnapshot` — instance *(class: `Worker`)*
@@ -3690,6 +3691,7 @@ Total: 2817 entries across 115 modules.
 - `postMessageToThread` — module
 - `receiveMessageOnPort` — module
 - `ref` — instance *(class: `Worker`)*
+- `removeEventListener` — instance
 - `setEnvironmentData` — module
 - `startCpuProfile` — instance *(class: `Worker`)*
 - `startHeapProfile` — instance *(class: `Worker`)*
